Use local static wkhtmltopdf library
Add a makefile target that clones the wkhtmltopdf
repository if it does not exist in lib/wkhtmltopdf.
This takes a lot of time, so libwkhtmltox.a can be added
in the root dir to prevent building the library.
The library could also be linked if already present on disk.

The wkhtmltox library is built and generates a static
library .a file.

Change mdpdf to link against the local static
wkhtmltox library.